* Added new random number generation to all modules. * Upgraded types in all modules to use quint64/32. * Fixed sillyness with deletes. * Added some scaffolding for fixed point in powers/roots modules. * Division works, but assumes integer result. Still need to add a checkbox   and make it an option. * Many things to fix and validate before it's ready with decimals, but works   fairly well at this point.

+++ powersconfigframe.cpp
@@ -21,22 +21,22 @@ void PowersConfigFrame::setModule(PowersModule *mod)
     this->module = mod;
 }
 
-void PowersConfigFrame::setMinimum(int min)
+void PowersConfigFrame::setMinimum(quint64 min)
 {
     ui->minNumberLineEdit->setText(QString("%1").arg(min));
 }
 
-void PowersConfigFrame::setMaximum(int max)
+void PowersConfigFrame::setMaximum(quint64 max)
 {
     ui->maxNumberLineEdit->setText(QString("%1").arg(max));
 }
 
-void PowersConfigFrame::setPowerMinimum(int min)
+void PowersConfigFrame::setPowerMinimum(quint32 min)
 {
     ui->minPowerLineEdit->setText(QString("%1").arg(min));
 }
 
-void PowersConfigFrame::setPowerMaximum(int max)
+void PowersConfigFrame::setPowerMaximum(quint32 max)
 {
     ui->maxPowerLineEdit->setText(QString("%1").arg(max));
 }
